+++ frontend/src/component/AutoTurning.jsx
@@ -1,16 +1,16 @@
-// 页面动画效果和节拍器
+// 定时翻页
 import React, { Component } from "react";
-import { NodeIndexOutlined } from "@ant-design/icons";
+import { FieldTimeOutlined } from "@ant-design/icons";
 
-class Metronome extends Component {
+class AutoTurning extends Component {
     render() {
         return (
             <div className={`${"tool-icon"} ${"toolIcon"}`}>
                 {/* "TODO: change this icon" */}
-                <NodeIndexOutlined className={`${"inOutlinedIcon"}`} />
+                <FieldTimeOutlined className={`${"inOutlinedIcon"}`} />
             </div>
         );
     }
 }
 
-export default Metronome;
+export default AutoTurning;